Pos devenu float pour rien

+++ AbletonPos.pde
@@ -23,12 +23,12 @@ void abletonPos ()
 
       if (networkSize == 6 )
         {
-          posAverage = ( Pos[0] + Pos[1] + Pos[1] + Pos[2] + Pos[3] + Pos[4] + Pos[5] )
+          posAverage =(int) ( Pos[0] + Pos[1] + Pos[1] + Pos[2] + Pos[3] + Pos[4] + Pos[5] )
                                                              / networkSize;
          }
       if (networkSize == 10 )
         {
-          posAverage = ( Pos[0] + Pos[1] + Pos[1] + Pos[2] + Pos[3] + Pos[4] + Pos[5] + 
+          posAverage = (int)( Pos[0] + Pos[1] + Pos[1] + Pos[2] + Pos[3] + Pos[4] + Pos[5] + 
                          Pos[6] + Pos[7] + Pos[8] + Pos[9] ) / networkSize;
          }
